requirement.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.13
django-otp==1.3.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djongo==1.3.6
frozenlist==1.4.1
idna==3.6
kombu==5.3.6
multidict==6.0.5
phonenumbers==8.13.33
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pymongo==3.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
twilio==9.0.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
